refactor: login screen, form control components
