--- src/components/atoms/buttons/index.ts
@@ -1,2 +1,0 @@
-export {default as ButtonLink} from './ButtonLink';
-export {default as PrimaryButton} from './PrimaryButton';
