+++ src/components/atoms/form-controls/common/types.ts
@@ -1,0 +1,21 @@
+export type FormControlPropsType = {
+  label: string;
+  placeholder?: string;
+  value?: string;
+  defaultValue?: string;
+  onChange?: (value: string) => void;
+  disabled?: boolean;
+  // custom
+  hasError?: boolean;
+  flat?: Boolean;
+};
+
+export type FormControlKeyboardType =
+  | 'text'
+  | 'number'
+  | 'decimal'
+  | 'search'
+  | 'email'
+  | 'tel'
+  | 'url'
+  | 'chat';
