+++ src/components/atoms/form-controls/common/utils.ts
@@ -1,26 +1,5 @@
 import {InputModeOptions, KeyboardTypeOptions} from 'react-native';
-
-export type FormControlPropsType = {
-  label: string;
-  placeholder?: string;
-  value?: string;
-  defaultValue?: string;
-  onChange?: (value: string) => void;
-  disabled?: boolean;
-  // custom
-  hasError?: boolean;
-  flat?: Boolean;
-};
-
-export type FormControlKeyboardType =
-  | 'text'
-  | 'number'
-  | 'decimal'
-  | 'search'
-  | 'email'
-  | 'tel'
-  | 'url'
-  | 'chat';
+import {FormControlKeyboardType} from './types';
 
 export class FormControlUtil {
   static getInputMode = (type: FormControlKeyboardType) =>
